Add animated code example
