+++ packages/docs/app/layout.tsx
@@ -2,10 +2,18 @@ import { RootProvider } from "fumadocs-ui/provider";
 import "./global.css";
 import { Inter } from "next/font/google";
 import type { ReactNode } from "react";
+import { Metadata, MetadataRoute, ResolvingMetadata } from "next";
 
 const inter = Inter({
   subsets: ["latin"],
 });
+
+export const metadata: Metadata = {
+  title: {
+    template: "Next-Yak | %s",
+    default: "Next-Yak",
+  },
+};
 
 export default function Layout({ children }: { children: ReactNode }) {
   return (
