+++ pnpm-workspace.yaml
@@ -50,6 +50,7 @@ catalogs:
     react-dom: 19.0.0
     react-resizable-panels: 2.1.7
     shiki: 1.24.4
+    shiki-magic-move: 0.5.2
     styled-components: 6.1.13
     swc-loader: 0.2.6
     tailwindcss: 3.4.17
